+++ insanic/models.py
@@ -77,7 +77,7 @@ class RequestService:
         else:
             service_type = self.request_service
 
-        return ",".join([service_type, self.destination_service, self.source_ip, self.destination_service])
+        return f"{service_type} - {self.destination_service}:{self.destination_version}({self.source_ip})"
 
     def __iter__(self):
         yield ("source", self.request_service)