+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog for insanic
 - FIX: gethostname on logs to get from cache
 - FIX: config can vault resolution from gethostbyaddr to gethostbyname
 - UPDATE: updates default backlog connnections to 65535
+- UPDATE: increase graceful shutdown timeout to 29
 
 
 0.8.2 (2019-10-16)